pyproject.toml
[tool.poetry]
name = "wildberries-project"
version = "0.1.0"
description = ""
authors = ["Saatarko <saatarko@tut.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
redis = "^5.1.1"
requests = "^2.32.3"
drf-spectacular = "^0.27.2"
beautifulsoup4 = "^4.12.3"
eventlet = "^0.37.0"
kombu = "^5.4.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
